+++ src/versioned/26.2/java/org/lavro/carpetlir/LIRSettings.java
@@ -47,12 +47,12 @@ public class LIRSettings {
     public static boolean boneMealGrassifyDirt = false;
 
     @Rule(
-            categories = {"LIR", "FEATURE", "RENEWABLE"}
+            categories = {"LIR", "FEATURE"}
     )
     public static boolean obsidianHardnessReinforcedDeepslate = false;
 
     @Rule(
-            categories = {"LIR", "FEATURE", "RENEWABLE"}
+            categories = {"LIR", "FEATURE"}
     )
     public static boolean silkTouchableReinforcedDeepslate = false;
 
@@ -62,7 +62,7 @@ public class LIRSettings {
     public static boolean wardensDropReinforcedDeepslate = false;
 
     @Rule(
-            categories = {"LIR", "FEATURE", "RENEWABLE"}
+            categories = {"LIR", "FEATURE"}
     )
     public static boolean pistonHarvestableAmethysts = false;
 }